requirements.txt
flask-bootstrap=3.3.7.1=py_0
flask-compress=1.5.0=py_0
flask-login=0.4.1=py37_0
flask-sqlalchemy=2.4.0=py_0
flask-wtf=0.14.2=py37_1
jinja2=2.10.3=py_0
joblib=0.13.2=py37_0
jpeg=9b=hb83a4c4_2
json5=0.8.5=py_0
jsonschema=3.0.2=py37_0
jupyter=1.0.0=py37_7
jupyter_client=5.3.3=py37_1
jupyter_console=6.0.0=py37_0
jupyter_core=4.5.0=py_0
jupyterlab=1.1.4=pyhf63ae98_0
jupyterlab_server=1.0.6=py_0
kealib=1.4.12=h07cbb95_0
keyring=18.0.0=py37_0
kiwisolver=1.1.0=py37ha925a31_0
krb5=1.16.1=hc04afaa_7
lazy-object-proxy=1.4.2=py37he774522_0
libarchive=3.3.3=h0643e63_5
libboost=1.67.0=hd9e427e_4
libcurl=7.65.3=h2a8f88b_0
libgdal=2.3.3=h10f50ba_0
libiconv=1.15=h1df5818_7
libkml=1.3.0=he5f2a48_4
liblief=0.9.0=ha925a31_2
libnetcdf=4.6.1=h411e497_2
libpng=1.6.37=h2a8f88b_0
libpq=11.2=h3235a2c_0
libsodium=1.0.16=h9d3ae62_0
libspatialindex=1.9.3=h33f27b4_0
libspatialite=4.3.0a=hc36aec2_19
libssh2=1.8.2=h7a1dbc1_0
libtiff=4.0.10=hb898794_2
libxml2=2.9.9=h464c3ec_0
libxslt=1.1.33=h579f668_0
llvmlite=0.29.0=py37ha925a31_0
locket=0.2.0=py37_1
lxml=4.4.1=py37h1350720_0
lz4-c=1.8.1.2=h2fa13f4_0
lzo=2.10=h6df0209_2
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=1.1.1=py37he774522_0
matplotlib=3.1.1=py37hc8f65d3_0
mccabe=0.6.1=py37_1
menuinst=1.4.16=py37he774522_0
mistune=0.8.4=py37he774522_0
mkl=2019.4=245
mkl-service=2.3.0=py37hb782905_0
mkl_fft=1.0.14=py37h14836fe_0
mkl_random=1.1.0=py37h675688f_0
mock=3.0.5=py37_0
more-itertools=7.2.0=py37_0
mpl-scatter-density=0.6=py_0
mpmath=1.1.0=py37_0
msgpack-python=0.6.1=py37h74a9793_1
msys2-conda-epoch=20160418=1
multipledispatch=0.6.0=py37_0
munch=2.5.0=py_0
navigator-updater=0.2.1=py37_0
nbconvert=5.6.0=py37_1
nbformat=4.4.0=py37_0
networkx=2.3=py_0
nltk=3.4.5=py37_0
nose=1.3.7=py37_2
notebook=6.0.1=py37_0
numba=0.45.1=py37hf9181ef_0
numexpr=2.7.0=py37hdce8814_0
numpy=1.16.5=py37h19fb1c0_0
numpy-base=1.16.5=py37hc3f5095_0
numpydoc=0.9.1=py_0
olefile=0.46=py37_0
openpyxl=3.0.0=py_0
openssl=1.1.1d=he774522_2
packaging=19.2=py_0
pandas=0.25.1=py37ha925a31_0


